+++ src/config/types.tsx
@@ -1,10 +1,3 @@
 export type ClientRouteType = {
     [key: string]: string | ClientRouteType;
 };
-
-export type RegisterFormData = {
-    id: string;
-    password: string;
-    confirmPassword: string;
-    email: string;
-};
