+++ src/types/auth.d.ts
@@ -1,0 +1,6 @@
+export interface RegisterFormData {
+    id: string;
+    password: string;
+    confirmPassword: string;
+    email: string;
+}
